refactor: port-adapter 패턴 적용
