+++ src/main/java/backend/sumnail/domain/nail_shop_hashtag/repository/NailShopHashtagRepositoryImpl.java
@@ -1,6 +1,7 @@
 package backend.sumnail.domain.nail_shop_hashtag.repository;
 
 import backend.sumnail.domain.nail_shop_hashtag.entity.NailShopHashtag;
+import backend.sumnail.domain.nail_shop_hashtag.service.port.NailShopHashtagRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
