+++ src/main/java/backend/sumnail/domain/recentsearch/repository/RecentSearchRepositoryImpl.java
@@ -1,6 +1,7 @@
 package backend.sumnail.domain.recentsearch.repository;
 
 import backend.sumnail.domain.recentsearch.entity.RecentSearch;
+import backend.sumnail.domain.recentsearch.service.port.RecentSearchRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
