+++ src/main/java/backend/sumnail/domain/station/service/StationService.java
@@ -2,7 +2,7 @@ package backend.sumnail.domain.station.service;
 
 import backend.sumnail.domain.station.controller.dto.response.StationFindAllResponse;
 import backend.sumnail.domain.station.entity.Station;
-import backend.sumnail.domain.station.repository.StationRepository;
+import backend.sumnail.domain.station.service.port.StationRepository;
 import java.util.List;
 import lombok.Builder;
 import lombok.RequiredArgsConstructor;
