+++ src/main/java/backend/sumnail/domain/user/service/port/UserRepository.java
@@ -1,4 +1,4 @@
-package backend.sumnail.domain.user.repository;
+package backend.sumnail.domain.user.service.port;
 
 import backend.sumnail.domain.user.entity.User;
 import java.util.Optional;
