+++ src/main/java/backend/sumnail/domain/user_nail_shop/repository/UserNailShopRepositoryImpl.java
@@ -3,6 +3,7 @@ package backend.sumnail.domain.user_nail_shop.repository;
 import backend.sumnail.domain.nail_shop.entity.NailShop;
 import backend.sumnail.domain.user.entity.User;
 import backend.sumnail.domain.user_nail_shop.entity.UserNailShop;
+import backend.sumnail.domain.user_nail_shop.service.port.UserNailShopRepository;
 import java.util.List;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
