+++ src/main/java/backend/sumnail/global/config/WebSecurityConfig.java
@@ -1,6 +1,6 @@
 package backend.sumnail.global.config;
 
-import backend.sumnail.domain.user.repository.UserRepository;
+import backend.sumnail.domain.user.service.port.UserRepository;
 import backend.sumnail.global.config.jwt.CustomAccessDeniedHandler;
 import backend.sumnail.global.config.jwt.CustomAuthenticationEntryPoint;
 import backend.sumnail.global.config.jwt.JwtAuthorizationFilter;
