+++ src/test/java/backend/sumnail/mock/FakeHashtagRepository.java
@@ -1,7 +1,7 @@
 package backend.sumnail.mock;
 
 import backend.sumnail.domain.hashtag.entity.Hashtag;
-import backend.sumnail.domain.hashtag.repository.HashtagRepository;
+import backend.sumnail.domain.hashtag.service.port.HashtagRepository;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
