+++ src/test/java/backend/sumnail/mock/FakeStationRepository.java
@@ -1,7 +1,7 @@
 package backend.sumnail.mock;
 
 import backend.sumnail.domain.station.entity.Station;
-import backend.sumnail.domain.station.repository.StationRepository;
+import backend.sumnail.domain.station.service.port.StationRepository;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
