+++ src/test/java/backend/sumnail/mock/FakeUserNailShopRepository.java
@@ -3,7 +3,7 @@ package backend.sumnail.mock;
 import backend.sumnail.domain.nail_shop.entity.NailShop;
 import backend.sumnail.domain.user.entity.User;
 import backend.sumnail.domain.user_nail_shop.entity.UserNailShop;
-import backend.sumnail.domain.user_nail_shop.repository.UserNailShopRepository;
+import backend.sumnail.domain.user_nail_shop.service.port.UserNailShopRepository;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
